fix(api): mark read-only fields as readOnly (#12735)
## Motivation

Described in https://github.com/kumahq/kuma/issues/12734. This is only
part of the work, the part needed for the TF provider. We should have
validation not only there but on the API as well.

## Implementation information

Add "readOnly" parameter to generated OAPI schema. For modification /
creation time do it always. For `status` do it only when it's present in
the schema.

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

xrel https://github.com/kumahq/kuma/issues/12734

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>