feat(kuma-cp): inbound name should be taken from pod instead of service (#12783)
## Motivation

At this moment we are taking inbound name from service port name. In
order to get rid of kuma.io/service we need to take name from pod port
name. Since we are not using name at the moment, it is safe to switch
this.

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>