ci(lint): check if policies have properly defined properties (#12794)
The linter should be complete. The only policy fully converted is
MeshAccessLog. The other policies have just targetRef/to/from modified.

Left to do in next PRs:
- support resources (should be straightforward, there is nothing
specific to policies here)
- fix linting errors

## Motivation

As mentioned in the issue, we identified 4 types of resources and their
combination of default/pointer/omitempty that will play nicely with TF.

## Implementation information

Implement a linter using tools/go/analysis and adjust the resources.

## Supporting documentation

Part of https://github.com/kumahq/kuma/issues/12790

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>