feat(api-server): add support for label filters (#12840)
## Motivation

We now have labels on all our entities. Some of these labels are very
useful (kuma.io/zone for example).
We should be able to use it when filtering api responses.

## Implementation information

Support a basic subset of https://kong-aip.netlify.app/aip/160/ only for
labels.

This works for all resource types and while a little costly it should
suffice for GUI and kumactl use cases

## Supporting documentation

Fix #12809

Signed-off-by: Charly Molter <charly.molter@konghq.com>