fix(api-server): add defaulter in MeshService and MeshMultizoneService
There are fields like targetPort on port name that default to another
value if not set.
This can't be expressed with annotations so defaulting with a
`Defaulter` is usefull in these cases.

There's an open question around what this would do with things like TF
as the PUT and GET return something different.

Signed-off-by: Charly Molter <charly.molter@konghq.com>