fix(kds): add all valid resources and skip only not valid (backport #12776) (#12855)
## Motivation

While working on
[kumahq/kuma#12719](https://github.com/kumahq/kuma/issues/12719), I
noticed some strange behavior. If there are X resources and one is
invalid, none of the valid resources are added. Once the invalid
resource is fixed, the previously valid resources are still not added.

We send a NACK for the invalid resource type, but there's no way to
reject only a single resource of that type—[we have to NACK the entire
resource type](https://github.com/envoyproxy/envoy/issues/5739).
Additionally, since the hash in the cache remains unchanged for valid
resources, they are not resent (this is how Delta xDS works), and only
the last changed resource is sent again.

## Implementation information
* Changed the logic to mark IsInitialRequest when there is a NACK, since
Delta XDS, on the second request, sends only the diff and not all
resources, even if the first request was NACKed.
* Added an error log to display NACK, making it easier to identify
incorrect resources.
* Filtered out invalid resources and stored valid ones, while still
sending a NACK with the invalid resources so the user is aware that some
are incorrect.

backport #12776

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>