feat(kds): don't delete zone created Secret (#12768)
## Motivation

The user might have a CertManager running in the zone or require a
secret that is accessible only within a specific zone, particularly for
a MeshGateway certificate that can be created in the zone and reference
a secret. Currently, if a user creates a certificate in the zone and
then restarts the control plane or creates another secret in the global
control plane, the zone detects the presence of a secret that shouldn’t
exist there, as secrets are not allowed in the zone.

## Implementation information

* Allow creating secrets in a zone without syncing them to the global
control plane.
* Handle name conflicts: If a user tries to create a secret in the
global control plane with the same name as an existing secret in a
specific zone, the system attempts to store it. If a
ResourceAlreadyExists error occurs, the resource is skipped, and a NACK
is returned.
* Introduced nackError in KDS sync to return a NACK when a
ResourceAlreadyExists error is encountered in the database.
* Added the SkipConflictResource option, enabled only for `Secrets`.
* Implemented E2E and unit tests to validate the changes.

Question:
Do we need `kuma.io/origin` validation?

## Supporting documentation

Needs https://github.com/kumahq/kuma/pull/12776
Fix https://github.com/kumahq/kuma/issues/12719

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>