chore(gha): added a workflow to trigger build on master branch (#12897)
## Motivation

We should trigger `build-test-distribute.yaml` manually instead of
pushing a commit directly into the PR. This approach increases the
visibility of the state of CI in our tools.

## Implementation information

Added a new workflow that triggers build-test-distribute.yaml and then
checks if it runs successfully. If it starts, the job completes. I kept
the ci-stability job so we can use it to check PR change stability.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/11624

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>