feat(kuma-cp): add support for inspect api for new kind Dataplane and section name for selecting single inbound (#12644)
Fix: #12361

## Motivation

We have introduced new kind and we need inspect api support for this, as
it turns out it works out of the box so I've just added test for it.

This needs https://github.com/kumahq/kuma/pull/12573 to be merged first

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>