fix(meshpassthrough): disable tls and http inspector for mysql protocol (#12839)
## Motivation

When trying to communicate with MySQL, the user wants to enable
communication with the database. Unfortunately, the MySQL protocol is
slightly different, and the usual method of configuring filter chains
does not work when using MeshPassthrough.

## Implementation information

* Added `original_dst` listener filter
* Added a protocol `mysql` which works only with CIDR/IP and requires
port since we need to disable `tls_inspector` and `http_inspector`
listener filters for the port
* Mysql protocol creates a `tcp_proxy` but with with disabled listener
filters
* Added test

> Why added mysql protocol and not just disable on ports with TCP
protocol?

The user may have rules with HTTP traffic and TCP traffic on the same
port matching different IP. example: tls matching on port 8080(IP:
192.168.1.1) and TCP matching on 8080 (IP: 172.1.1.1), that would
disable TLS inspector on the port 8080 and wouldn't match

## Supporting documentation


https://dev.mysql.com/doc/dev/mysql-server/8.4.3/page_protocol_connection_phase_packets.html
https://github.com/envoyproxy/envoy/issues/21044

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>