refactor(dp-server): remove metadata tracker (#12922)
## Motivation

We were tracking the metadata in its own dataplane callbacks.
In practice this metadata was only used inside the sync_tracker which is
itself a dataplane callback.
With https://github.com/kumahq/kuma/pull/12886 we only have 1 connection
at a time for a dataplane key so there's no need to manage metadata
changes.

## Implementation information

We now pass the metadata directly inside the sync_tracker when creating
the watchdog.
This makes the code more streamlined and removes a hard to understand
abstraction that added little value.

## Supporting documentation

Part of #12881

Signed-off-by: Charly Molter <charly.molter@konghq.com>