test(e2e): rewrite leader election with postgres (#12943)
## Motivation

While analyzing the flake, I noticed that our test assumes the first
instance is always the leader. This is mostly true, but in some cases,
high CPU usage or network failures can cause the leader to be lost. In
the flake we observed, the first instance was initially the leader but
missed a heartbeat, causing leadership to switch to another instance and
resulting in a failed assertion.

leader1
```
2025-02-25T05:23:54.685Z	ERROR	postgres-leader	heartbeat missed: lock is already released	{"error": "lock is already released"}
2025-02-25T05:23:54.685Z	INFO	bootstrap	leader lost
```
leader2
```
2025-02-25T05:24:00.125Z	INFO	bootstrap	leader acquired
```

## Implementation information

Changed the logic to assume any of these 2 instances can be a leader and
store information about the leader and follower.

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>