feat(kuma-cp): allow `sectionName` and `labels` in `targetRef` (#11819)
## Motivation

According to https://github.com/kumahq/kuma/issues/11572 there's no real
reason to have such validation.
However, I didn't get if this part from the issue `If one of the
matching meshService doesn't have such a sectionName it should simply
not match just like it's already the case for labels.` already works in
this way or it needs to be implemented.

## Implementation information

The change is pretty simple.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/11572

---------

Signed-off-by: Nikolai M <ne9zzz@gmail.com>
Signed-off-by: Ilya Lobkov <ilya.lobkov@konghq.com>
Co-authored-by: Ilya Lobkov <ilya.lobkov@konghq.com>