fix(kds): rework cross zone syncing (#12893)
# Motivation

Up until now it was really hard to define the behaviour of resources in
KDS in a generic way leading to code complexity and bugs:

- Resources that only existed in global could be created on zones
- Problems with secrets colliding between global and zone
- Partial sync (only some secrets synced from Global to zones)

# Implementation details

We rethink KDS using behaviors instead on a per policy basis.
The behaviors are:
- Can a resource be created on global or zone
- Is a resource be "consumed" (used) on global or zone
- Is a resource synced across zones

With this we can decide what gets sent around in KDS the filters and
mapper.

What's important is that these flags do not decide *which* resources are
synced these are further rules in KDS for .e.g:
- Don't send to global resources that originated from global
- Don't send to zone resources coming from this same zone
- Only send across policies that are of "producer" style (which makes
resources like MeshService work out the box)
- Don't sync some configs and secrets that are specific to a CP (tokens
private keys for .e.g).

# Supporting documentation

Fix #11740
Created #12916

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>