chore(api): export all computed labels to be used in another project (#13056)
## Motivation

**This PR does not change anything from `kuma` perspective.**

Terraform will show changes to labels that are automatically generated,
like `kuma.io/mesh` so applying:

```
resource "konnect_mesh_external_service" "my_meshexternalservice" {
  cp_id = konnect_mesh_control_plane.my_meshcontrolplane.id

  labels = {
    environment = "production"
    # "kuma.io/mesh" = konnect_mesh.default.name
  }

...
}
```

Will succeed but on the next `apply` we would see `kuma.io/mesh` because
it's automatically generated:

```
  # konnect_mesh_external_service.my_meshexternalservice will be updated in-place
  ~ resource "konnect_mesh_external_service" "my_meshexternalservice" {
      ~ labels            = {
          - "kuma.io/mesh" = "mesh-9" -> null
            # (1 unchanged element hidden)
        }
        name              = "external-service-example"
        # (7 unchanged attributes hidden)
    }

Plan: 0 to add, 1 to change, 0 to destroy.
```

This PR allows the provider to import `AllComputedLabels` and remove
them from terraform plan.

## Implementation information

Export AllComputedLabels and make sure that in the future this list is
updated.

This can be achieved without this PR but the labels might get out of
sync in the future if we don't update it in the provider.

## Supporting documentation

Xrel: https://github.com/Kong/platform-api/issues/1023

Signed-off-by: slonka <slonka@users.noreply.github.com>