test(unit): compare file content instead of yaml (#13105)
## Motivation

While debugging a flake I noticed that `MatchGoldenYAML` doesn't work
correctly. In the match, it compares only the first element of the file.

https://pkg.go.dev/gopkg.in/yaml.v3#section-readme
> Multi-document unmarshalling is not yet implemented, and base-60
floats from YAML 1.1 are purposefully not supported since they're a poor
design and are gone in YAML 1.2.

## Implementation information

Changed to `MatchGoldenEqual` since we are using it in other tests.
Also, change sort to `strings.Cmp`

> Changelog: skip

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>