feat(kumactl): skip not system namespace when exporting federation (#13114)
## Motivation

We don't want to export policies to a namespace other than the system
namespace for the `federation-with-policies` profile, as they cannot be
applied globally.

## Implementation information

If the resource's namespace is different from the system namespace, skip
it. If the namespace is empty, it means the resource is cluster-scoped,
so include it.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13110

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>