test(e2e): improve universal app run (#13033)
## Motivation

It was very hard to have a complete and readable state of things in the
test report because we were executing one of ssh cmds in different
places.
We now keep a persistent ssh session for the whole time of the app which
makes things easier to track and the output easier to consume.

## Implementation information

the ssh package was completely redesigned to keep an ssh session rather
than forking to execute `ssh` commands. This makes things easier to
follow, easier to write lengthy scripts without caring about escaping
things correctly for bash.
Once this was done we could start a CP or run a DP with a single cmdLine
instead of a list of complex calls.

This makes the report a lot easier as it avoids a lot of noisy stuff.

Other minor changes:
- Stop doing retries inside the framework. Either we should retry
because it's setup or simply in an eventually. Also stop using the retry
from terratest as it's very verbose.
- ignore for dockerfiles were including things that were not necessary,
this made the building of images longer.
- Once this change was done we could run into clock skews with tokens.
So moved the token issuance date a few seconds in the past

## Supporting documentation

Fix #13123

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>