chore(e2e): rewrite e2e config init (#13165)
## Motivation

That way in a parent project we can get the default and overwrite only
the parts of the config that we care about. So if a change happens like
in
https://github.com/kumahq/kuma/pull/12982/files#diff-4097df65e73d6b3a2813df0ea3c87f15a6c43ba3aca9b73b3e60346758ef0344R65
we won't need to modify the parent project.

Instead of https://github.com/kumahq/kuma/pull/13163 we can just
manually init the e2e config and in parent project use `init` to change
the `Config`.

First review parent project.

## Implementation information

Manually init the e2e configuration by moving the init function to
`RunE2ESpecs`.

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>