fix(kds): detect properly hanging stream for the same zone (#12983)
## Motivation

In a multi-zone architecture, we run a global control plane and multiple
zone control planes. Each zone can have multiple instances running, but
only the leader is responsible for communicating with the global control
plane.

Currently, if there is an issue with a zone or a proxied stream that
does not properly detect the connection closure, we might end up in a
situation where a zone control plane restart goes unnoticed. This could
result in multiple open streams from a single zone to the global control
plane. Since the health check only verifies the zone name and tenant ID,
a newly connected instance of the control plane makes all previous
health check detectors pass.

This issue occurs when a gRPC client and server, both configured with
Keepalive, communicate through a Gateway that is also configured with
the gRPC protocol. We can reproduce it with the following setup:
```
Zone → Gateway (gRPC) → Global
```
By introducing packet loss on the Zone, we can observe the following log
messages in the control plane:

`keepalive ping failed to receive ACK within timeout`
However, despite this failure, the connection remains open in the
control plane, leading to unnecessary resource allocation. This suggests
that the Gateway might be misconfigured. In such cases, we should
implement safeguards to prevent hanging connections.

## Implementation information

We could track the last connection time by recording it in an event and
later checking if the same client attempts to reconnect. If a
reconnection occurs, we cancel the old stream.

To handle scenarios where multiple instances of the global control plane
are running, we will store this information in zoneInsight.
Additionally, we will use a periodic verification mechanism (zone
watcher) to ensure the recorded time remains accurate.

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>