perf(rules): add `withNegation` flag to simplify `to` policy flow (#13151)
> [!NOTE]
> This is not a problem when using `meshServices.mode: Exclusive` with
policies in a new style

When a user has multiple `to` policies for a single top-level target
reference and a default to Mesh target, CPU usage spikes due to the
expensive matching process.
<img width="2053" alt="Screenshot 2025-03-20 at 18 43 36"
src="https://github.com/user-attachments/assets/259af8fb-357f-4cfe-ae54-257ce0cd36d9"
/>

After analyzing the issue with @lobkovilya, we discovered that for `to`
policies, our logic checks every permutation of tags. However, this is
unnecessary because the only possible tags for `to` policies are
`kuma.io/service`, or no tags at all when the target is `Mesh`.

The only exception is `MeshHTTPRoute`, where we generate the
`__rule-matches-hash__` tag, which points to a hash of the matcher.
Since we don’t need to evaluate all permutations, we decided to iterate
through the subsets and match them directly with the policies. To
optimize further, we first deduplicate entries to avoid redundant
computations.

After implementing these changes, I deployed the update and ran a
performance profile, which showed significant improvements.

![image](https://github.com/user-attachments/assets/db30fc2e-aafe-4e16-a8ef-df8a9401bd25)

fix: https://github.com/kumahq/kuma/issues/13149

---------

Signed-off-by: Ilya Lobkov <ilya.lobkov@konghq.com>
Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>
Co-authored-by: Ilya Lobkov <ilya.lobkov@konghq.com>