test(e2e): fix getting cp logs after panic (#13278)
## Motivation

When there is a panic in the logs but the test suite succeeds, the
control plane logs are not included in the debug package, making it more
difficult to debug the issue.

## Implementation information

The issue was that when an assertion fails at [this
line](https://github.com/kumahq/kuma/blob/master/test/framework/debug.go#L27),
it stops the resource dumping process. To prevent this, we need to use
GinkgoRecover to ensure the function continues executing. More details
can be found in the [Ginkgo
documentation](https://onsi.github.io/ginkgo/#mental-model-how-ginkgo-handles-failure).

Also, we were missing a dump log step on universal

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13260

> Changelog: skip

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>