feat(dns): embedded DNS server instead of coreDNS (#13124)
## Motivation

Add an embedded DNS server that resolves mesh local hostnames.
This feature is disabled by default and needs to be enabled with
`kuma_dns_proxy_port ` on DPs in universal to a non 0 value or simply
with `kuma_runtime_kubernetes_injector_builtin_dns_experimental_proxy
=true` on CP in Kubernetes (this leverages sidecar injection).

## Implementation information

We leveraged the same system as MeshMetric. The DNS map is exposed in
the `_kuma:dynamicconfig` listener, we poll this from the DP at a set
frequency and reload the data.

We improved the whole configfetcher to make it reusable across different
components. We also setup the endpoint so that it works with [etag
caching](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/ETag)
while avoids unnecessary reloading and processing, we can therefore
refresh the configs more quickly as well.

For the moment I added a new matrix for e2e test. If this is stable I'll
switch the default.

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>