fix(kuma-dp): change log level of fetcher to debug (#13399)
## Motivation

`kuma-dp` starts a fetcher to retrieve the configuration for various
components, such as DNS and mesh metrics. For DNS, this is controlled by
a flag that enables or disables it. The issue lies with mesh metrics,
which are dynamically configured via policy. This means we need to
attempt to fetch the configuration on each tick. If no configuration is
found, we log a message. However, this log is overly aggressive and can
obscure other potential issues.

## Implementation information

Change log level to debug

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13357

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>