fix(kumactl): change mtls backend builtin to provided and skipvalidation (#13318)
## Motivation

When a user exports a mesh with the mTLS type set to builtin, it is
exported as is. However, this is incorrect when the user attempts to
import it.

During import, this mesh should be treated as provided since the secrets
are already included. Additionally, applying a builtin mesh triggers the
generation of new secrets, even though existing ones are already
available.

## Implementation information

Simplified the logic for better readability:
* Ordered resources: `Mesh` → `Secrets` → ... → `GlobalSecrets`
* Changed mesh from builtin to provided and added `skipValidate` to
ensure `Secret` creation can happen after the mesh.
* Enabled federation test.
* Fixed Kubernetes export, as `skipCreatingInitialPolicies` was not
being set for the mesh.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13109
Fix: https://github.com/kumahq/kuma/issues/13217

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>