test(framework): allow specify watch namespaces for the KIC installer (#13421)
## Motivation

Fixing a bug in the KIC installer of the test framework which prevent us
from specifying watch namespaces.

The issue was introduced by this PR
https://github.com/kumahq/kuma/pull/12848 which tried to isolate the
impact between multiple test suites on the same cluster.

The issue is causing some integration test failures in child-repo
projects.

## Implementation information

In our test framework, add a new option `watchNamespaces` onto the
`k8sDeployment` struct of the KIC installer. In this way, we
differenciate the semantic of `ingressNamespace` and `watchNamespaces`
in that struct.

The mentioned PR above made the KIC installer include `watchNamespaces`
parameter and made its value to equal `ingressNamespace`. This design
prevented the installed KIC to watch Kong Gateway resources/plugins from
any other namespaces than the namespace in which the KIC components are
placed in. It's a major shortage to the original implmentation.

## Supporting documentation


https://docs.konghq.com/kubernetes-ingress-controller/2.8.x/references/cli-arguments/


> Changelog: skip

<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Jay Chen <1180092+jijiechen@users.noreply.github.com>