test(e2e): wait for stat to show up for MeshRateLimit (#13424)
## Motivation

Noticed a flake

## Implementation information

While checking the logs, I noticed that the metrics were present in the
dump, but during the test, they couldn't be found. This might be caused
by a delay in configuration delivery, as we only check once if the
metrics are available. I added an Eventually block to wait until the
metrics appear, which should help avoid issues due to config propagation
delays.

> Changelog: skip

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>