fix(secrets): correctly compare other meshes secrets when using crossMesh (#13426)
## Motivation

While investigating test failures, I noticed a flake that caused the
secret not to be loaded.

## Implementation information

While reviewing the code, I found that we were only comparing the number
of meshes, but not verifying whether a specific mesh still exists. I
added a change to detect cases where a mesh was removed and another was
added in a short period of time. I also added a test and a metric to
track mesh generation changes.

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>