fix(dns): enable e2e tests with embedded dns and reduce TTL (#13457)
## Motivation

We want to switch to embedded DNS but we had some flakes.

## Implementation information

I've noticed that we fail in the following scenario:

1. We create an ExternalService, which allocates 240.0.0.3 for a real
domain (something we don't do for MES).
2. Kong Gateway makes a request to the domain and receives a response
with 240.0.0.3.
3. We update the DNS configuration and remove this DNS entry (since
we're testing passthrough and removed the ExternalService).
4. Kong Gateway has cached the IP, and we no longer see logs indicating
requests to the DNS server.

The default TTL is 3600s, so the cached record isn't refreshed.
This creates a potential issue when a service is removed and later
redeployed with a new VIP address. Due to the long TTL, Kong might
continue using the stale IP. A shorter TTL would help mitigate this
risk.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13353

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>
Co-authored-by: Charly Molter <charly@molter.io>