feat(meshhttproute): rename `routeConfiguration` and `routes` to use KRI (#13322)
## Motivation

Adopting KRI format for routes. Need this PR to support targeting routes
with policies.

## Implementation information

* when outbound listener is generated for the real MeshService, the
routes are going to be named with KRI
* rename `BackendRefOriginIndex` to `OriginByMatches` since we actually
need to have this mapping even when `default.backendRefs` is empty (i.e.
when MeshHTTPRoute has only `default.filters`
* refactor `SortRules` function, because it was doing a lot of things
and it was annoying to fix test for it
* refactor test cases that had to construct BackendRefOriginIndex
manually, now `Origin` and `OriginByMatches` are automatically set in
tests

---------

Signed-off-by: Ilya Lobkov <ilya.lobkov@konghq.com>