feat(kuma-cp): add metadata field to DataplaneInsight for storing XDS metadata (#13522)
## Motivation

On bootstrap we send a bunch of metadata (for .e.g feature flags) we
should be able to persist these in `DataplaneInsight` this would enable
reliable future proof implementation of things like:
https://github.com/kumahq/kuma-gui/issues/ 2472

## Implementation information

This adds a new `metadata` field of type `google.protobuf.Struct` to the
`DataplaneInsight` proto. The field is intended to store arbitrary xDS
metadata associated with a dataplane. Updates the `.proto`, generated
`.pb.go` code, and the status sink logic to support this addition.

## Supporting documentation

Closes https://github.com/kumahq/kuma/issues/ 13490

---------

Signed-off-by: Bart Smykla <bartek@smykla.com>