fix(e2e): remove dnsLookupFamily from e2e envoy config test response (#13582)
## Motivation

Since we run e2e tests on both ipv4 and ipv6, `dnsLookupFamily` is
different. In ipv4 `dnslookupFamily` is set to V4_ONLY,
and in the case of ipv6 this field is default, so it is missing in the
config. Which causes e2e tests to fail on ipv6 in some cases

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>