feat(api): rename zone-ingress/global-secrets to k8s conventions name (#13418)
## Motivation

While working on the missing ingress config dump, I noticed a lack of
consistency in resource names. The naming differs between inspect and
get. We should be consistent, especially since we previously made a
change for this purpose, but this resource wasn't covered at the time.

## Implementation information

Introduced a new field in the description that allows users to provide
an alias for the path/kumactl command. We then create endpoints for the
alias and also include it in the kumactl command.

It shouldn't break the old kumactl since we have 2 endpoints
`zone-ingresses` and `zoneingresses`, but the new client does requests
to `zoneingresses`.

Other ideas:
Should we return a 301 when someone calls zone-ingresses instead of
registering two endpoints?

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13359

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>