fix(kds): don't override status field (#13618)
## Motivation

In a multizone setup using MeshService mode, we synchronize objects
across zones. Each object contains a status field, which is
intentionally removed during synchronization on the global control
plane. This allows a remote zone to allocate a VIP and hostname for the
MeshService based on its local configuration.

However, when the connection to the global control plane is
reestablished, the system detects a change in the resource and overrides
the object in the zone, even when there are no actual changes. This
happens because the global control plane sync a version of the
MeshService with an empty status field, while the zone contains an
updated status. As a result, the objects appear to differ and trigger an
unnecessary update.

The override of MeshService status field was casuing multiple IP
allocations and could cause more frequent updates of XDS snapshots.

## Implementation information

First of all, we don't need to compare the status field of synchronized
resources, since only the zone runs a generator that sets the status,
while the global control plane always sync the empty status field.
Additionally, we should ensure that the status field is removed from all
resources when syncing them from the global. Since the status shouldn't
change, we should set the status of the existing object instead of
overriding it.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/13592

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>